delete now obsolete WaitingList table
